[BUG]: Conflict로 인한 Dependency 수정
