+++ GaeManda/Projects/Core/GMDExtensions/UIViewController+Extensions.swift
@@ -7,25 +7,10 @@
 //
 
 import UIKit
-import GMDUtils
 
 public extension UIViewController {
 	override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
 		super.touchesBegan(touches, with: event)
 		self.view.endEditing(true)
-	}
-}
-
-public extension UIViewController {
-	func hideTabBar() {
-		guard let parent = navigationController?.parent as? FloatingTabBarPresentable else { return }
-
-		parent.dismissTabBar()
-	}
-	
-	func showTabBar() {
-		guard let parent = navigationController?.parent as? FloatingTabBarPresentable else { return }
-		
-		parent.presentTabBar()
 	}
 }
